refactor: :building_construction: フォロー関連の処理を切り出し
