+++ lib/core/common_widget/button/follow_or_unfollow_button.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../feature/user_profile/application/user_follow_service.dart';
-import '../../../feature/user_profile/application/user_profile_state.dart';
+import '../../../feature/user_follow/application/user_follow_service.dart';
+import '../../../feature/user_follow/application/user_follow_state.dart';
 import '../shimmer_widget.dart';
 import 'filled_button.dart';
 import 'outlined_button.dart';
