+++ lib/feature/auth/application/auth_service.dart
@@ -10,10 +10,10 @@ import '../../definition/repository/write_definition_repository.dart';
 import '../../user_config/application/user_config_state.dart';
 import '../../user_config/repository/device_info_repository.dart';
 import '../../user_config/repository/user_config_repository.dart';
+import '../../user_follow/repository/user_follow_repository.dart';
 import '../../user_profile/application/user_id_list_state.dart';
 import '../../user_profile/domain/user_profile.dart';
 import '../../user_profile/repository/storage_repository.dart';
-import '../../user_profile/repository/user_follow_repository.dart';
 import '../../user_profile/repository/user_profile_repository.dart';
 import '../../user_profile/util/user_list_type.dart';
 import '../repository/auth_repository.dart';
