+++ lib/feature/user_follow/presentation/page/following_and_follower_list/following_and_follower_list_page.dart
@@ -7,9 +7,10 @@ import '../../../../../core/common_widget/simple_widget_for_empty.dart';
 import '../../../../../core/common_widget/stickey_tab_bar_deligate.dart';
 import '../../../../../util/extension/scroll_controller_extension.dart';
 import '../../../../auth/application/auth_state.dart';
-import '../../../application/user_profile_state.dart';
-import '../../../util/user_list_type.dart';
-import '../../component/profile_list.dart';
+import '../../../../user_profile/application/user_profile_state.dart';
+import '../../../../user_profile/presentation/component/profile_list.dart';
+import '../../../../user_profile/util/user_list_type.dart';
+import '../../../application/user_follow_state.dart';
 
 @RoutePage()
 class FollowingAndFollowerListPage extends ConsumerWidget {
