+++ lib/feature/user_profile/application/user_profile_state.dart
@@ -1,10 +1,7 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../auth/application/auth_state.dart';
 import '../../user_profile/repository/user_profile_repository.dart';
-import '../domain/follow_count.dart';
 import '../domain/user_profile.dart';
-import '../repository/user_follow_repository.dart';
 
 part 'user_profile_state.g.dart';
 
@@ -26,34 +23,4 @@ Future<String?> userIdSearchByPublicId(
       .searchUserProfileByPublicId(publicId);
 
   return userProfileDoc?.id;
-}
-
-@riverpod
-Future<FollowCount> followCount(
-  FollowCountRef ref,
-  String userId,
-) async {
-  final userFollowCountDoc =
-      await ref.read(userFollowRepositoryProvider).fetchUserFollowCount(userId);
-
-  return FollowCount.fromDocument(userFollowCountDoc);
-}
-
-@riverpod
-Future<bool> isFollowing(IsFollowingRef ref, String targetUserId) async {
-  final currentUserId = ref.read(userIdProvider)!;
-
-  return ref
-      .read(userFollowRepositoryProvider)
-      .isFollowing(currentUserId, targetUserId);
-}
-
-@riverpod
-Future<List<String>> followingIdList(
-  FollowingIdListRef ref,
-  String targetUserId,
-) async {
-  return ref
-      .read(userFollowRepositoryProvider)
-      .fetchAllFollowingIdList(targetUserId);
 }
