+++ lib/feature/user_profile/presentation/component/profile_tile.dart
@@ -3,13 +3,13 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../../core/common_widget/adaptive_overflow_text.dart';
-import '../../../../../core/common_widget/avatar_network_image_widget.dart';
-import '../../../../../core/common_widget/error_and_retry_widget.dart';
-import '../../../../../core/router/app_router.dart';
-import '../../../../../util/exception/database_exception.dart';
-import '../../../../../util/logger.dart';
-import '../../../application/user_profile_state.dart';
+import '../../../../core/common_widget/adaptive_overflow_text.dart';
+import '../../../../core/common_widget/avatar_network_image_widget.dart';
+import '../../../../core/common_widget/error_and_retry_widget.dart';
+import '../../../../core/router/app_router.dart';
+import '../../../../util/exception/database_exception.dart';
+import '../../../../util/logger.dart';
+import '../../application/user_profile_state.dart';
 import 'profile_tile_shimmer.dart';
 
 class ProfileTile extends ConsumerWidget {
