+++ test/feature/user_profile/application/user_follow_service_test.mocks.dart
@@ -8,12 +8,12 @@ import 'dart:async' as _i6;
 import 'package:cloud_firestore/cloud_firestore.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:riverpod_annotation/riverpod_annotation.dart' as _i8;
+import 'package:teigi_app/feature/user_follow/repository/entity/user_follow_count_document.dart'
+    as _i3;
+import 'package:teigi_app/feature/user_follow/repository/user_follow_repository.dart'
+    as _i5;
 import 'package:teigi_app/feature/user_profile/domain/user_id_list_state.dart'
     as _i4;
-import 'package:teigi_app/feature/user_profile/repository/entity/user_follow_count_document.dart'
-    as _i3;
-import 'package:teigi_app/feature/user_profile/repository/user_follow_repository.dart'
-    as _i5;
 
 import 'user_follow_service_test.dart' as _i7;
