+++ test/feature/user_profile/application/user_profile_state_test.mocks.dart
@@ -7,15 +7,15 @@ import 'dart:async' as _i7;
 
 import 'package:cloud_firestore/cloud_firestore.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:teigi_app/feature/user_follow/repository/entity/user_follow_count_document.dart'
+    as _i4;
+import 'package:teigi_app/feature/user_follow/repository/user_follow_repository.dart'
+    as _i9;
 import 'package:teigi_app/feature/user_profile/domain/user_id_list_state.dart'
     as _i5;
 import 'package:teigi_app/feature/user_profile/domain/user_profile.dart' as _i8;
-import 'package:teigi_app/feature/user_profile/repository/entity/user_follow_count_document.dart'
-    as _i4;
 import 'package:teigi_app/feature/user_profile/repository/entity/user_profile_document.dart'
     as _i3;
-import 'package:teigi_app/feature/user_profile/repository/user_follow_repository.dart'
-    as _i9;
 import 'package:teigi_app/feature/user_profile/repository/user_profile_repository.dart'
     as _i6;
